Replace huge advance value by a safer one with the same behavior
The `default.json` font file contains a font advance value of `1e400`.
This value is too big to be represented as-is even by IEEE 754 64-bit
floating point numbers as used by Java and, by extension, Minecraft.
Gson, the JSON deserialization library used by Minecraft, cannot
properly represent such a number either, but the deserialization process
is lenient and chugs along with a practically close enough value.

According to the ECMA JSON standard, numbers in JSON may be arbitrarily
big, so such a number is technically within specification. However, the
standard highlights the possibility that different programs disagree on
the way they represent numbers, which in practice can cause
interoperability issues.

One such interoperability issue happens with PackSquash, which relies on
`serde_json` for JSON parsing, and that parser (reasonbly) refuses to
deal with numbers that cannot be represented as floats.

In order to avoid interoperability problems, let's use an advance value
of `1e37`, which can be represented even by 32-bit floats and works
equally well enough in practice as an "infinite" value.